Create README.md
Our program works through a Chat screen, it could be an application on phones or a link in a health website. The program tries to diagnose the symptoms suffered by the user. The engage with the user by asking him about the kind of symptoms he had, after that the program tray guess the type of disease that the user could have and suggests appropriate treatment for him, or suggested booking an appointment with the doctor if the problem could not be solved.